refactor: split pkg to model
most of the code is for model, let's split it more in detail
